done with debugging

+++ lab2/DebugExercise/DebugExercise1.java
@@ -6,8 +6,8 @@ package DebugExercise;
  */
 public class DebugExercise1 {
     public static int divideThenRound(int top, int bottom) {
-        int quotient = top / bottom;
-        int result = Math.round(quotient);
+        double quotient = (double)top / (double)bottom;
+        int result = (int)Math.round(quotient);
         return result;
     }
 
@@ -17,7 +17,7 @@ public class DebugExercise1 {
         int result = divideThenRound(t, b);
         System.out.println("round(" + t + "/" + b + ")=" + result);
 
-        int t2 = 9;
+        int t2 =9;
         int b2 = 4;
         int result2 = divideThenRound(t2, b2);
         System.out.println("round(" + t2 + "/" + b2 + ")=" + result2);
